Changed log time format to RFC3339
