+++ logs.go
@@ -2,9 +2,11 @@ package goss
 
 import (
 	"fmt"
+	"io"
 	"log"
 	"os"
 	"strings"
+	"time"
 
 	"github.com/goss-org/goss/util"
 	"github.com/hashicorp/logutils"
@@ -16,7 +18,8 @@ func setLogLevel(c *util.Config) error {
 		MinLevel: logutils.LogLevel("WARN"),
 		Writer:   os.Stderr,
 	}
-	log.SetOutput(filter)
+	log.SetFlags(0) // Turn off standard timestamp flags
+	log.SetOutput(&timestampedWriter{filter})
 	for _, lvl := range filter.Levels {
 		cLvl := strings.ToUpper(c.LogLevel)
 		if string(lvl) == cLvl {
@@ -26,4 +29,13 @@ func setLogLevel(c *util.Config) error {
 		}
 	}
 	return fmt.Errorf("Unsupported log level: %s", c.LogLevel)
+}
+
+type timestampedWriter struct {
+	wrappedWriter io.Writer
+}
+
+func (t *timestampedWriter) Write(b []byte) (int, error) {
+	timestamp := time.Now().UTC().Format(time.RFC3339)
+	return fmt.Fprintf(t.wrappedWriter, "%s %s", timestamp, b)
 }
